requirements.txt
absl-py==0.11.0
aiocontextvars==0.2.2
appnope==0.1.0
astor==0.8.1
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
backcall==0.2.0
batchgenerators==0.21
bleach==3.1.5
cached-property==1.5.2
certifi==2020.12.5
charset-normalizer==2.0.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dicom2nifti==2.2.12
docopt==0.6.2
efficientnet-pytorch==0.6.3
einops==0.3.0
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==3.2
imagecodecs==2020.5.30
imageio==2.9.0
imgviz==1.4.0
immutables==0.16
importlib-metadata==1.6.1
iniconfig==1.1.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
linecache2==1.0.0
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
MedPy==0.4.0
mistune==0.8.4
ml-collections==0.1.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
nibabel==3.2.1
nnunet==1.6.4
notebook==6.0.3
numpy==1.19.5
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pipreq==0.4
pipreqs==0.4.10
pluggy==1.0.0
pretrainedmodels==0.7.4
prettyprint==0.1.5
prettytable==2.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.14.0
ptyprocess==0.6.0
py==1.10.0
pycodestyle==2.6.0
pydicom==2.1.2
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.5
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
segmentation-models-pytorch==0.2.0
self-attention-cv==1.2.3
Send2Trash==1.5.0
SimpleITK==2.0.2
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorboardX==2.4
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.4.12
toml==0.10.1
torch==1.10.0
torchmetrics==0.5.0
torchsummary==1.5.1
torchvision==0.11.1
tornado==6.0.4
tqdm==4.56.0
traceback2==1.4.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.10.0.2
unittest2==1.1.0
urllib3==1.26.6
vit-pytorch==0.20.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
yarg==0.1.9
zipp==3.1.0
